Data Analytics
Finding answers to our Need:
Workspace Access Control is only available in the Databricks Operational Security Package. Learn moreThis notebook will be published publicly and anyone with the link can view it
Need:
Is there a pattern to monthly variations in the ratings of businesses? Comparing reviews to the average for that business, or each month to the prior month, do users rate businesses higher or lower than the average in certain months? Is this more of a seasonal or monthly issue? Does this vary geographically or by other sub-category?